refactor: delete unused files, unify others
