+++ dash/src/components/ScrollHint.tsx
