--- dash/src/lib/sorting/sortPos.ts
@@ -1,5 +1,0 @@
-type PosObject = { position: string };
-
-export const sortPos = (a: PosObject, b: PosObject) => {
-	return parseInt(a.position) - parseInt(b.position);
-};
