--- dash/src/types/history.type.ts
@@ -1,7 +1,0 @@
-export type DriverHistory = {
-	nr: number;
-	gapAhead: string[];
-	gapLeader: string[];
-	laptime: string[];
-	sectors: [string[], string[], string[]];
-};
