--- dash/src/types/recap.type.ts
@@ -1,6 +1,0 @@
-import { State } from "./state.type";
-
-export type Snapshot = {
-	state: State;
-	timestamp: string;
-};
